Change the algorithm for movement and collision
The previous implementation stored X and Y position as 8.8 fixed point
numbers and added 8.8 fixed point deltas to them every frame.  While this
seems natural, it makes implementing collisions difficult since we end up
having to calculate fractional Y-movement on X-collision, and vice versa.
That would require relatively slow fixed-point calculations, which I'd
like to avoid.  It was possible to avoid the calculations in the special
case where the ball only collides with the edges of the rectangular field,
but paddle-ball collisions are more complicated.

Therefore, the new strategy is to store the X and Y coordinates of the ball
as simple 8-bit unsigned integers.  Instead of X and Y velocities, we use
countdown timers that increment/decrement X or Y when they reach zero.  These
timers count "ticks" of which there are currently 16 per frame.  Once a timer
reaches zero, it is reset to a preset value.  The higher the preset value,
the slower the motion.  The preset value is effectively proportional to the
reciprocal of the velocity.

This strategy allows us to easily step the movement pixel-by-pixel in order
to process collisions with pixel perfect accuracy, and without involving any
fixed-point math.